[patch] [file data] fix

+++ patches/april_2013/p07_update_file_data_2.py
@@ -6,6 +6,9 @@ def execute():
 	for doctype in webnotes.conn.sql_list("""select table_name from `information_schema`.`columns`
 		where table_schema=%s and column_name='file_list'""", webnotes.conn.cur_db_name):
 			doctype = doctype[3:]
+			
+			if not webnotes.conn.exists("DocType", doctype): continue
+			
 			update_file_list(doctype, singles)
 			
 			webnotes.conn.sql("""delete from `tabCustom Field` where fieldname='file_list'
